Fix several bugs in Android and Linux * Fix invalid exponential moving average in Counter (Linux) * Make printing network switcher state tidy (Linux) * Add skipping verbose messages in some classes (Android)   - ClientAdapter   - MainActivity.ReceivingThread * Fix crash error in sending control message (Android) * Fix network switcher's state checking (Android) * Add handling EAGAIN in ServerSockets (Linux)   - RfcommServerSocket   - TcpServerSocket   - ServerSocket * Add handling kernel I/O buffer full (Linux) * Add 60s waiting after trace runner finishes (Linux)

+++ linux/inc/NetworkSwitcher.h
@@ -111,14 +111,14 @@ protected:
 class Stats {
 public:
   /* Statistics used to print present status */
-  int ema_queue_arrival_speed = 0;
+  float ema_queue_arrival_speed = 0;
 
   /* Statistics used in CoolSpots Policy */
   int now_total_bandwidth = 0;
 
   /* Statistics used in Energy-aware & Latency-aware Policy */
-  int ema_send_request_size = 0;
-  int ema_arrival_time_us = 0;
+  float ema_send_request_size = 0;
+  float ema_arrival_time_us = 0;
   int now_queue_data_size = 0;
 };
